mock data for CodeSandbox
