+++ package.json
@@ -71,7 +71,8 @@
     "lint": "eslint \"src/**/*.{js,ts,tsx}\" --fix --max-warnings=0",
     "client": "webpack serve --mode=development",
     "server": "node server/server.js",
-    "start": "concurrently \"npm run server\" \"npm run client\"",
+    "start:dev": "concurrently \"npm run server\" \"npm run client\"",
+    "start": "webpack serve --mode=production",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
     "test": "jest",
