+++ src/features/telegramChart/store/reducers/chartData.ts
@@ -1,4 +1,5 @@
 import {dataChart} from '@/api/telegramChart';
+import {mockDataCharts} from '@/api/telegramChart/mockData';
 import {TChartData} from '@/api/telegramChart/types';
 import {createAsyncThunk, createSlice} from '@reduxjs/toolkit';
 import {CHART_DATA_REDUCER, MODULE_NAME} from '../../constants';
@@ -6,8 +7,11 @@ import {CHART_DATA_REDUCER, MODULE_NAME} from '../../constants';
 export const fetchDataChart = createAsyncThunk(
   'data_charts/getDataChart',
   async () => {
-    const response = await dataChart.get();
-    return (await response.json()) as TChartData;
+    if (process.env.NODE_ENV === 'development') {
+      const response = await dataChart.get();
+      return (await response.json()) as TChartData;
+    }
+    return mockDataCharts;
   }
 );
